% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile-likelihood.R
\name{binomial_loglik}
\alias{binomial_loglik}
\title{Binomial Log-Likelihood Helper}
\usage{
binomial_loglik(x, n, p)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of trials}

\item{p}{Probability of success}
}
\value{
Log-likelihood value
}
\description{
Computes the binomial log-likelihood for a single proportion.
}
\keyword{internal}
