% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_sve.R
\name{profile_likelihood_ci_single}
\alias{profile_likelihood_ci_single}
\title{Profile Likelihood Confidence Interval for SVE (Single Observation)}
\usage{
profile_likelihood_ci_single(x0, x1, n0, n1, level = 0.95)
}
\arguments{
\item{x0}{Number of events in unvaccinated group (scalar)}

\item{x1}{Number of events in vaccinated group (scalar)}

\item{n0}{Sample size of unvaccinated group (scalar)}

\item{n1}{Sample size of vaccinated group (scalar)}

\item{level}{Confidence level}
}
\value{
List with estimate and CI bounds
}
\description{
Profile Likelihood Confidence Interval for SVE (Single Observation)
}
\keyword{internal}
