% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial_data.R
\docType{data}
\name{sim_trial_data}
\alias{sim_trial_data}
\title{Simulated Vaccine Trial Data}
\format{
\subsection{\code{sim_trial_data}}{

A data frame with 500 rows and 60 columns:
\describe{
\item{id}{ID}
\item{vaccination}{Vaccine status (1 = vaccinated, 0 = unvaccinated)}
\item{age}{Age}
\item{baseline_risk}{Baseline risk (low, medium, or high)}
\item{time}{Survival time}
\item{status}{Censoring status}
}
}
}
\usage{
sim_trial_data
}
\description{
A simulated data set.
}
\keyword{datasets}
