% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sve_from_effect.R
\name{sve_effect}
\alias{sve_effect}
\title{Compute SVE from Effect Measure}
\usage{
sve_effect(theta)
}
\arguments{
\item{theta}{Relative effect measure}
}
\value{
Symmetric vaccine efficacy
}
\description{
Compute SVE from Effect Measure
}
\keyword{internal}
