% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile-likelihood.R
\name{sve_p0_bounds}
\alias{sve_p0_bounds}
\title{Determine Valid p0 Range for SVE Constraint}
\usage{
sve_p0_bounds(sve_target, tol = 1e-08)
}
\arguments{
\item{sve_target}{Target SVE value}

\item{tol}{Tolerance for bounds (default 1e-8)}
}
\value{
Two-element vector with lower and upper bounds for p0
}
\description{
Determine Valid p0 Range for SVE Constraint
}
\keyword{internal}
