% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile-likelihood.R
\name{sve_profile_ci_single}
\alias{sve_profile_ci_single}
\title{Profile Likelihood Confidence Interval for SVE (Single Observation)}
\usage{
sve_profile_ci_single(x0, x1, n0, n1, level = 0.95, correction = TRUE)
}
\arguments{
\item{x0}{Number of events in unvaccinated group (scalar)}

\item{x1}{Number of events in vaccinated group (scalar)}

\item{n0}{Sample size of unvaccinated group (scalar)}

\item{n1}{Sample size of vaccinated group (scalar)}

\item{level}{Confidence level}

\item{correction}{Logical. Whether to perform a bias correction, default: \code{TRUE}.}
}
\value{
List with estimate and CI bounds
}
\description{
Computes profile likelihood CI for a single set of proportions.
}
\keyword{internal}
