% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile-likelihood.R
\name{sve_to_p1}
\alias{sve_to_p1}
\title{Convert SVE to Proportions Given p0}
\usage{
sve_to_p1(p0, sve_target)
}
\arguments{
\item{p0}{Proportion in group 0}

\item{sve_target}{Target SVE value}
}
\value{
Corresponding p1 value
}
\description{
Given p0 and a target SVE, solve for p1.
}
\keyword{internal}
