% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile-likelihood.R
\name{sve_to_theta}
\alias{sve_to_theta}
\title{Convert SVE to Theta (Relative Risk)}
\usage{
sve_to_theta(sve)
}
\arguments{
\item{sve}{Symmetric vaccine efficacy value}
}
\value{
Corresponding theta value
}
\description{
Convert SVE to Theta (Relative Risk)
}
\keyword{internal}
