% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_sve.R
\name{sve_var}
\alias{sve_var}
\title{Compute Variance of SVE via Delta Method}
\usage{
sve_var(p0, p1, n0, n1)
}
\arguments{
\item{p0}{Proportion in unvaccinated group}

\item{p1}{Proportion in vaccinated group}

\item{n0}{Sample size of unvaccinated group}

\item{n1}{Sample size of vaccinated group}
}
\value{
Variance of SVE
}
\description{
Compute Variance of SVE via Delta Method
}
\keyword{internal}
