% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sve_from_effect.R
\name{sve_var_effect}
\alias{sve_var_effect}
\title{Compute Variance of SVE via Delta Method}
\usage{
sve_var_effect(theta, var_log_theta, level)
}
\arguments{
\item{theta}{Relative effect measure}

\item{var_log_theta}{Variance of log(theta)}
}
\value{
Variance of SVE
}
\description{
Compute Variance of SVE via Delta Method
}
\keyword{internal}
