% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-profile-likelihood.R
\name{two_sample_binomial_loglik}
\alias{two_sample_binomial_loglik}
\title{Two-Sample Binomial Log-Likelihood}
\usage{
two_sample_binomial_loglik(x0, x1, n0, n1, p0, p1)
}
\arguments{
\item{x0}{Number of events in group 0}

\item{x1}{Number of events in group 1}

\item{n0}{Sample size of group 0}

\item{n1}{Sample size of group 1}

\item{p0}{Probability in group 0}

\item{p1}{Probability in group 1}
}
\value{
Log-likelihood value
}
\description{
Computes the log-likelihood for two independent binomial samples.
}
\keyword{internal}
